LayoutTests/fast/frames/srcdoc/srcdoc-can-be-in-qurks-mode.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
Normally srcdoc documents default to standards mode, but they can end up in
quirks mode with a sufficiently nutty DocType.
<iframe srcdoc='
  <!DOCTYPE banana>
  <script>
  alert(document.compatMode);
  </script>
'></iframe>
