LayoutTests/fast/frames/srcdoc/srcdoc-can-be-in-qurks-mode-expected.txt

ALERT: BackCompat
Normally srcdoc documents default to standards mode, but they can end up in quirks mode with a sufficiently nutty DocType.
